% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMean.R
\name{printMean}
\alias{printMean}
\title{Print mean values in a sentence}
\usage{
printMean(x)
}
\arguments{
\item{x}{A vector of numeric values of which to calculate the mean}
}
\value{
The output of \code{\link{print}} after calculating the \code{\link{mean}}
    and pasting it into a sentence
}
\description{
This function calculates the mean of its input and
    prints it into a sentence. The function is called in analysis.Rmd
}
\examples{
printMean(c(1,2,3))

}
